Integer.MAX_VALUE not Short.MAX_VALUE
git-svn-id: file:///home/toad/git-migration/temprepository/branches/db4o/freenet@21371 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/client/async/ClientRequestScheduler.java
@@ -551,7 +551,7 @@ public class ClientRequestScheduler implements RequestScheduler {
 		else if(PRIORITY_HARD.equals(choosenPriorityScheduler))
 			fuzz = 0;	
 		// schedCore juggles both
-		return schedCore.removeFirst(fuzz, random, offeredKeys, starter, schedTransient, transientOnly, notTransient, Short.MAX_VALUE, Short.MAX_VALUE, clientContext, container);
+		return schedCore.removeFirst(fuzz, random, offeredKeys, starter, schedTransient, transientOnly, notTransient, Short.MAX_VALUE, Integer.MAX_VALUE, clientContext, container);
 	}
 
 	public ChosenRequest getBetterNonPersistentRequest(ChosenRequest req) {
